+++ src/test/kotlin/org/jetbrains/intellij/IntelliJPluginSpec.kt
@@ -37,7 +37,8 @@ class IntelliJPluginSpec : IntelliJPluginSpecBase() {
                 IntelliJPluginConstants.RUN_PLUGIN_VERIFIER_TASK_NAME,
                 IntelliJPluginConstants.SETUP_DEPENDENCIES_TASK_NAME,
                 IntelliJPluginConstants.SIGN_PLUGIN_TASK_NAME,
-                IntelliJPluginConstants.VERIFY_PLUGIN_TASK_NAME
+                IntelliJPluginConstants.VERIFY_PLUGIN_TASK_NAME,
+                IntelliJPluginConstants.VERIFY_PLUGIN_CONFIGURATION_TASK_NAME
             ),
             tasks(IntelliJPluginConstants.GROUP_NAME),
         )