+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## next
 ### Added
-- Check if Java version is suitable for the current `intellij.platformVersion`
+- The [`verifyPluginConfiguration`](https://plugins.jetbrains.com/docs/intellij/tools-gradle-intellij-plugin.html#verifypluginconfiguration-tasl) task for validating the plugin project configuration.
 
 ### Changed
 - Change `IntelliJPluginConstants.ANDROID_STUDIO_PRODUCTS_RELEASES_URL` to `https://jb.gg/android-studio-releases-list.xml`